Move class Output to a different file
